Update so far...
Doesn't work yet - update tries to implement reading from SampleLiveMarketData, but has problems trying to communicate between different classes at the moment, as the classes aren't instantiated in OrderManager, but instead are connected through input/output streams. I will try to improve on this as soon as possible.

+++ javaproject/src/OrderClient/NewOrderSingle.java
@@ -5,10 +5,10 @@ import Ref.Instrument;
 
 public class NewOrderSingle implements Serializable {
 	public int size;
-	public float price;
+	public double price;
 	public Instrument instrument;
 
-	public NewOrderSingle(int size, float price, Instrument instrument) {
+	public NewOrderSingle(int size, double price, Instrument instrument) {
 		this.size = size;
 		this.price = price;
 		this.instrument = instrument;
